lottie: fix tvg::clamp template deduction on Xtensa
On Xtensa (ESP32), int32_t is defined as long rather than int.
This causes tvg::clamp<T> template deduction to fail when the first
argument is int32_t but the literal arguments 0/255 are int, as the
compiler cannot deduce a single T from mismatched types.

Cast all arguments to int32_t to ensure consistent template deduction
across all platforms.

+++ src/loaders/lottie/tvgLottieData.h
@@ -124,9 +124,9 @@ static inline RGB32 operator*(const RGB32& lhs, float rhs)
 static inline RGB32 lerp(const RGB32& s, const RGB32& e, float t)
 {
     return {
-        tvg::clamp((int32_t)(s.r + (e.r - s.r) * t), 0, 255),
-        tvg::clamp((int32_t)(s.g + (e.g - s.g) * t), 0, 255),
-        tvg::clamp((int32_t)(s.b + (e.b - s.b) * t), 0, 255)
+        tvg::clamp((int32_t)(s.r + (e.r - s.r) * t), (int32_t)0, (int32_t)255),
+        tvg::clamp((int32_t)(s.g + (e.g - s.g) * t), (int32_t)0, (int32_t)255),
+        tvg::clamp((int32_t)(s.b + (e.b - s.b) * t), (int32_t)0, (int32_t)255)
     };
 }
